refactor(client): move CSS domain tests to consistent location
- Move CSS domain tests from css/__tests__/ to __tests__/ for consistency
- Rename test files: css.test.ts, css-stylesheet.test.ts, css-utils.test.ts
- Update import paths to match new location
- All 57 tests passing
