+++ packages/client/src/cdp/domain/__tests__/css-utils.test.ts
@@ -6,7 +6,7 @@ import {
   toCssProperties,
   getLineCount,
   getLastLine,
-} from '../utils';
+} from '../css/utils';
 
 describe('CSS Utils', () => {
   describe('parseCssText', () => {
